refactor(control): split services.rs into per-service modules
Split the monolithic services.rs into a services/ directory with one file
per REAPI service:

- services/cas.rs — ContentAddressableStorage with tracing spans
- services/action_cache.rs — ActionCache with tracing spans
- services/capabilities.rs — Capabilities
- services/execution.rs — Execution with span propagation
- services/mod.rs — re-exports + shared pub(crate) digest helpers

Shared helpers (proto_to_digest, digest_to_proto) moved to mod.rs.
All RPC handlers now have tracing instrumentation.
